Simplify `RenameEntriesTransformer` to work only on one entry
